AbArticulus/requirements.txt
djangorestframework==2.3.12
django-extensions==1.3.3
flake8==2.2.1
flake8-print
flake8-debugger
nose==1.3.1
nose-progressive==1.5
django-nose==1.2
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
django-social-auth==0.7.28
